Suppress Redis Connection Errors
Previously, if the Store could not make a connection to Redis (e.g. if
the Redis server went down), it would prevent the request from
succeeding.  In most cases, however, an outage like this would be
temporary.  In that case, requests should continue to succeed, working
against the in-memory instance of the session for as long as possible.
This change ensures that these connection errors are well logged, but
that requests will succeed until such time as the server comes back
up.

[#64605700]