Java 6 Support
Previously, the session managers were all written using Java 7 syntax
and compile to Java 7 compatible byte-code.  Users requested that the
session manager work on Java 6 as well as Java 7 and Java 8.  This
change modifies the source syntax, removing the Java 7 specific code.
It also modifies the Maven configuration such that the create binaries
are Java 6 compatible.