Handle non-standard passwords properly
Previously, there were a couple of bugs in how non-standard passwords
were handled.  In truth, one of the bugs was how passwords were
handled in the case of `redis://localhost:6379/0` which is the
**typical** case.  As well, we didn't properly handle the case where
we were unable to get a resource from the Jedis pool and then tried to
return it anyway.

This change fixes both bugs.

[#75030176]